requirements.txt
streamlit
pandas
matplotlib
numpy
pillow
qrcode
